Move JPA entities and repositiories
Move JPA entities and repositiories to a package for more better structure project
